+++ TombEngine/Scripting/Internal/ReservedScriptNames.h
@@ -450,31 +450,34 @@ constexpr char ScriptReserved_StaticSetSlot[]		= "SetSlot";
 constexpr char ScriptReserved_StaticSetSolid[]		= "SetSolid";
 constexpr char ScriptReserved_StaticShatter[]		= "Shatter";
 
-//Collision Class and Tables
-static constexpr char ScriptReserved_Collision[]				= "Collision";
-static constexpr char ScriptReserved_Probe[]					= "Probe";
+// =========
+// COLLISION
+// =========
 
-static constexpr char ScriptReserved_GetColPosition[]			= "GetPosition";
-static constexpr char ScriptReserved_GetColRoomNumber[]			= "GetRoomNumber";
-static constexpr char ScriptReserved_GetColRoomName[]			= "GetRoomNumber";
-static constexpr char ScriptReserved_GetColRoom[]				= "GetRoom";
-static constexpr char ScriptReserved_GetFloorHeight[]			= "GetFloorHeight";
-static constexpr char ScriptReserved_GetCeilingHeight[]			= "GetCeilingHeight";
-static constexpr char ScriptReserved_GetWaterSurfaceHeight[]	= "GetWaterSurfaceHeight";
-static constexpr char ScriptReserved_GetFloorNormal[]			= "GetFloorNormal";
-static constexpr char ScriptReserved_GetCeilingNormal[]			= "GetCeilingNormal";
-static constexpr char ScriptReserved_GetFloorMaterialType[]		= "GetFloorMaterialType";
-static constexpr char ScriptReserved_GetCeilingMaterialType[]	= "GetCeilingMaterialType";
+constexpr char ScriptReserved_Collision[]					= "Collision";
 
-static constexpr char ScriptReserved_IsSteepFloor[]				= "IsSteepFloor";
-static constexpr char ScriptReserved_IsSteepCeiling[]			= "IsSteepCeiling";
-static constexpr char ScriptReserved_IsWall[]					= "IsWall";
-static constexpr char ScriptReserved_IsInsideSolidGeometry[]	= "IsInsideSolidGeometry";
-static constexpr char ScriptReserved_IsClimbableWall[]			= "IsClimbableWall";
-static constexpr char ScriptReserved_IsMonkeySwing[]			= "IsMonkeySwing";
-static constexpr char ScriptReserved_IsDeathTile[]				= "IsDeathTile";
+constexpr char ScriptReserved_Probe[]						= "Probe";
+constexpr char ScriptReserved_ProbeGetCeilingHeight[]		= "GetCeilingHeight";
+constexpr char ScriptReserved_ProbeGetCeilingMaterialType[]	= "GetCeilingMaterialType";
+constexpr char ScriptReserved_ProbeGetCeilingNormal[]		= "GetCeilingNormal";
+constexpr char ScriptReserved_ProbeGetFloorHeight[]			= "GetFloorHeight";
+constexpr char ScriptReserved_ProbeGetFloorMaterialType[]	= "GetFloorMaterialType";
+constexpr char ScriptReserved_ProbeGetFloorNormal[]			= "GetFloorNormal";
+constexpr char ScriptReserved_ProbeGetPosition[]			= "GetPosition";
+constexpr char ScriptReserved_ProbeGetRoom[]				= "GetRoom";
+constexpr char ScriptReserved_ProbeGetRoomName[]			= "GetRoomName";
+constexpr char ScriptReserved_ProbeGetRoomNumber[]			= "GetRoomNumber";
+constexpr char ScriptReserved_ProbeGetWaterSurfaceHeight[]	= "GetWaterSurfaceHeight";
+constexpr char ScriptReserved_ProbeIsClimbableWall[]		= "IsClimbableWall";
+constexpr char ScriptReserved_ProbeIsDeathTile[]			= "IsDeathTile";
+constexpr char ScriptReserved_ProbeIsInsideSolidGeometry[]	= "IsInsideSolidGeometry";
+constexpr char ScriptReserved_ProbeIsMonkeySwing[]			= "IsMonkeySwing";
+constexpr char ScriptReserved_ProbeIsSteepCeiling[]			= "IsSteepCeiling";
+constexpr char ScriptReserved_ProbeIsSteepFloor[]			= "IsSteepFloor";
+constexpr char ScriptReserved_ProbeIsWall[]					= "IsWall";
 
-static constexpr char ScriptReserved_GetSectorMaterialType[]	= "MaterialType";
+constexpr char ScriptReserved_MaterialType[]				= "MaterialType";
+
 // =====
 // TYPES
 // =====